+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - On `ConvertStringMoveUtils`, added `addMoveNumberToSanSequence` adding move number on given san sequence like `1. e4 e5 2.Nf3` for engine pv showing.
+  (also added on `ChessGame.toNumberedSan`)
+- On `ConvertStringMoveUtils`, added `removeMoveNumberFromSanSequence` removing move number on given san sequence like 
+  `1. e4 e5 2. Nf3 Nc6` to `e4 e5 Nf3 Nc6`.
+  (also added on `ChessGame.removeNumberFromSan`)
 
 ### Changed
 - On `UCIEngineWrapper`, changed `sanPv` string to add move number like `1. e4 e5 2. Nf3 Nc6`.